Cleaning up MongoDB support - moving out EventStore
