--- Source/Bifrost.MongoDb/EventClassMap.cs
@@ -1,18 +1,0 @@
-﻿/*---------------------------------------------------------------------------------------------
- *  Copyright (c) 2008-2017 Dolittle. All rights reserved.
- *  Licensed under the MIT License. See LICENSE in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-using Bifrost.Events;
-using MongoDB.Bson.Serialization;
-
-namespace Bifrost.MongoDb
-{
-    public class EventClassMap : BsonClassMap<IEvent>
-    {
-        public EventClassMap()
-        {
-            AutoMap();
-            
-        }
-    }
-}
