--- Source/Bifrost.MongoDb/Events/EventStore.cs
@@ -1,146 +1,0 @@
-﻿/*---------------------------------------------------------------------------------------------
- *  Copyright (c) 2008-2017 Dolittle. All rights reserved.
- *  Licensed under the MIT License. See LICENSE in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Bifrost.Events;
-using MongoDB.Bson;
-using MongoDB.Bson.Serialization;
-using MongoDB.Driver;
-
-namespace Bifrost.MongoDb.Events
-{
-	public class EventStore : IEventStore
-	{
-		const string CollectionName = "Events";
-		const string IncrementalKeysCollectionName = "_IncrementalKeys";
-		const string EventType = "EventType";
-		const string Generation = "Generation";
-		const string Version = "Version";
-		const string LogicalEventType = "LogicalEventType";
-		const string CurrentKey = "CurrentKey";
-
-		EventStorageConfiguration _configuration;
-		MongoClient _server;
-		IMongoDatabase _database;
-		IMongoCollection<BsonDocument> _collection;
-		IMongoCollection<BsonDocument> _incrementalKeysCollection;
-		IEventMigrationHierarchyManager _eventMigrationHierarchyManager;
-
-		static EventStore()
-		{
-			BsonSerializer.RegisterSerializer(typeof(EventSourceVersion), new EventSourceVersionSerializer());
-		}
-
-		public EventStore(EventStorageConfiguration configuration, IEventMigrationHierarchyManager eventMigrationHierarchyManager)
-		{
-			_configuration = configuration;
-			_eventMigrationHierarchyManager = eventMigrationHierarchyManager;
-			Initialize();
-		}
-
-		void Initialize()
-		{
-			var s = MongoClientSettings.FromUrl(new MongoUrl(_configuration.Url));
-			if (_configuration.UseSSL)
-			{
-				s.UseSsl = true;
-				s.SslSettings = new SslSettings
-				{
-					EnabledSslProtocols = System.Security.Authentication.SslProtocols.Tls12,
-					CheckCertificateRevocation = false
-				};
-			}
-			_server = new MongoClient(s);
-			_database = _server.GetDatabase(_configuration.DefaultDatabase);
-			_collection = _database.GetCollection<BsonDocument>(CollectionName);
-			_incrementalKeysCollection = _database.GetCollection<BsonDocument>(IncrementalKeysCollectionName);
-		}
-
-		public CommittedEventStream GetFor(IEventSource eventSource)
-		{
-			var eventSourceType = eventSource.GetType();
-			var builder = Builders<BsonDocument>.Filter;
-			var filter = builder.Eq("EventSourceId", eventSource.EventSourceId) & builder.Eq("EventSource", eventSourceType.AssemblyQualifiedName);
-
-			var cursor = _collection.Find<BsonDocument>(filter);
-			var documents = cursor.ToList();
-			var events = ToEvents(documents);
-			var stream = new CommittedEventStream(eventSource.EventSourceId, events);
-			return stream;
-		}
-
-		public void Commit(IEnumerable<EventAndEnvelope> eventsAndEnvelopes)
-		{
-            foreach( var eventAndEnvelope in eventsAndEnvelopes )
-            { 
-				var eventDocument = eventAndEnvelope.ToBsonDocument();
-				AddMetaData(eventAndEnvelope.Event, eventDocument);
-				_collection.InsertOne(eventDocument);
-			}
-		}
-
-		public EventSourceVersion GetLastCommittedVersionFor(IEventSource eventSource)
-		{
-			var filter = Builders<BsonDocument>.Filter.Eq("EventSourceId", eventSource.EventSourceId);
-			var @event = _collection.Find<BsonDocument>(filter).SortBy(d => d.GetElement(Version)).FirstOrDefault();
-			if (@event == null)
-				return EventSourceVersion.Zero;
-
-			return EventSourceVersion.FromCombined(@event[Version].AsDouble);
-		}
-
-		int GetNextEventId()
-		{
-			var currentValue = 0;
-			var query = Builders<BsonDocument>.Filter.Eq("_id", CollectionName);
-			var result = _incrementalKeysCollection.FindOneAndUpdate(query, Builders<BsonDocument>.Update.Inc(CurrentKey, 1), new FindOneAndUpdateOptions<BsonDocument>() { ReturnDocument = ReturnDocument.After, IsUpsert = true });
-
-			//TODO: Is this really necessary? The above is an upsert?
-			if (result?.AsBsonDocument == null)
-			{
-				var eventsCurrentValue = new BsonDocument();
-				eventsCurrentValue["_id"] = CollectionName;
-				eventsCurrentValue[CurrentKey] = 1;
-				_incrementalKeysCollection.InsertOne(eventsCurrentValue);
-				currentValue = 1;
-			}
-			else
-				currentValue = result.AsBsonDocument[CurrentKey].AsInt32;
-			return currentValue;
-		}
-
-		void AddMetaData(IEvent @event, BsonDocument eventDocument)
-		{
-			var eventType = @event.GetType();
-			var logicalEventType = _eventMigrationHierarchyManager.GetLogicalTypeFor(eventType);
-			var migrationLevel = _eventMigrationHierarchyManager.GetCurrentGenerationFor(logicalEventType);
-			eventDocument[EventType] = string.Format("{0}, {1}", eventType.FullName, eventType.Assembly.GetName().Name);
-			eventDocument[LogicalEventType] = string.Format("{0}, {1}", logicalEventType.FullName, logicalEventType.Assembly.GetName().Name);
-			//eventDocument[Generation] = migrationLevel;
-		}
-
-		void RemoveMetaData(BsonDocument document)
-		{
-			document.Remove(EventType);
-			document.Remove(LogicalEventType);
-			document.Remove(Generation);
-		}
-
-		IEnumerable<EventAndEnvelope> ToEvents(IEnumerable<BsonDocument> documents)
-		{
-            var events = new List<EventAndEnvelope>();
-
-            foreach (var document in documents)
-			{
-				var eventType = Type.GetType(document[EventType].AsString);
-				RemoveMetaData(document);
-				var instance = BsonSerializer.Deserialize(document, eventType) as IEvent;
-                events.Add(new EventAndEnvelope(null, instance));
-            }
-			return events;
-		}
-	}
-}
