+++ Source/Bifrost.Read.MongoDb/Bifrost.MongoDB.nuspec
