+++ Source/Bifrost.Read.MongoDb/EntityContext.cs
@@ -9,14 +9,22 @@ using MongoDB.Driver;
 using MongoDB.Bson;
 using System.Reflection;
 
-namespace Bifrost.MongoDb
+namespace Bifrost.Read.MongoDB
 {
+    /// <summary>
+    /// Represents an implementation of <see cref="IEntityContext{T}"/> for MongoDB
+    /// </summary>
+    /// <typeparam name="T">Type of entity</typeparam>
 	public class EntityContext<T> : IEntityContext<T>
 	{
 		EntityContextConnection _connection;
 		string _collectionName;
 		IMongoCollection<T> _collection;
 
+        /// <summary>
+        /// Initializes a new instance of <see cref="EntityContext{T}"/>
+        /// </summary>
+        /// <param name="connection"></param>
 		public EntityContext(EntityContextConnection connection)
 		{
 			_connection = connection;
@@ -25,30 +33,35 @@ namespace Bifrost.MongoDb
 			_collection = _connection.Database.GetCollection<T>(_collectionName);
 		}
 
-
+        /// <inheritdoc/>
 		public IQueryable<T> Entities
 		{
 			get { return _collection.AsQueryable<T>(); }
 		}
 
+        /// <inheritdoc/>
 		public void Attach(T entity)
 		{
 		}
 
+        /// <inheritdoc/>
 		public void Insert(T entity)
 		{
 			_collection.InsertOne(entity);
 		}
 
+        /// <inheritdoc/>
 		public void Update(T entity)
 		{
 			Save(entity);
 		}
 
+        /// <inheritdoc/>
 		public void Delete(T entity)
 		{
 		}
 
+        /// <inheritdoc/>
 		public void Save(T entity)
 		{
 			var idProperty = GetIdProperty(entity);
@@ -57,10 +70,27 @@ namespace Bifrost.MongoDb
 			_collection.ReplaceOne(filter, entity, new UpdateOptions() { IsUpsert = true });
 		}
 
+        /// <inheritdoc/>
 		public void Commit()
 		{
 		}
 
+        /// <inheritdoc/>
+		public T GetById<TProperty>(TProperty id)
+        {
+            var objectId = GetObjectId(id);
+            return _collection.Find(Builders<T>.Filter.Eq("_id", objectId)).FirstOrDefault();
+        }
+
+        /// <inheritdoc/>
+		public void DeleteById<TProperty>(TProperty id)
+        {
+            var objectId = GetObjectId(id);
+            _collection.DeleteOne(Builders<T>.Filter.Eq("id", objectId));
+        }
+
+
+        /// <inheritdoc/>
 		public void Dispose()
 		{
 		}
@@ -71,11 +101,6 @@ namespace Bifrost.MongoDb
 		}
 
 
-		public T GetById<TProperty>(TProperty id)
-		{
-			var objectId = GetObjectId(id);
-			return _collection.Find(Builders<T>.Filter.Eq("_id", objectId)).FirstOrDefault();
-		}
 
 		BsonValue GetObjectId<TProperty>(TProperty id)
 		{
@@ -85,13 +110,6 @@ namespace Bifrost.MongoDb
 
 			var idAsValue = BsonValue.Create(idValue);
 			return idAsValue;
-		}
-
-
-		public void DeleteById<TProperty>(TProperty id)
-		{
-			var objectId = GetObjectId(id);
-			_collection.DeleteOne(Builders<T>.Filter.Eq("id", objectId));
 		}
 	}
 }
