+++ Source/Bifrost.Read.MongoDb/EntityContextConnection.cs
@@ -4,17 +4,21 @@
  *--------------------------------------------------------------------------------------------*/
 using Bifrost.Entities;
 using Bifrost.Execution;
-using Bifrost.MongoDb.Concepts;
+using Bifrost.Read.MongoDB.Concepts;
 using MongoDB.Bson.Serialization;
 using MongoDB.Driver;
 
-namespace Bifrost.MongoDb
+namespace Bifrost.Read.MongoDB
 {
+    /// <summary>
+    /// Represents an implementation of <see cref="IEntityContextConnection"/> for MongoDB
+    /// </summary>
     public class EntityContextConnection : IEntityContextConnection
     {
-        public MongoClient Server { get; private set; }
-        public IMongoDatabase Database { get; private set; }
-
+        /// <summary>
+        /// Initializes a new instance of <see cref="EntityContextConnection"/>
+        /// </summary>
+        /// <param name="configuration"></param>
         public EntityContextConnection(EntityContextConfiguration configuration)
         {
             var s = MongoClientSettings.FromUrl(new MongoUrl(configuration.Url));
@@ -34,6 +38,20 @@ namespace Bifrost.MongoDb
             BsonSerializer.RegisterSerializationProvider(new ConceptSerializationProvider());
         }
 
+        /// <summary>
+        /// Gets the <see cref="MongoClient"/> representing the connection to the server
+        /// </summary>
+        public MongoClient Server { get; }
+
+        /// <summary>
+        /// Gets the <see cref="IMongoDatabase"/> - a reference to the actual database
+        /// </summary>
+        public IMongoDatabase Database { get; }
+
+        /// <summary>
+        /// Initialize the connection
+        /// </summary>
+        /// <param name="container"></param>
         public void Initialize(IContainer container)
         {
