+++ Source/Bifrost.Read.MongoDb/Properties/AssemblyInfo.cs
