+++ Source/Bifrost.Read.MongoDb/packages.config
