+++ Source/Bifrost.Read.MongoDb/project.json
