+++ Source/Solutions/Bifrost_All.sln
@@ -19,8 +19,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.JSON", "..\Bifrost.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.Ninject", "..\Bifrost.Ninject\Bifrost.Ninject.csproj", "{3EE03158-1BE0-470C-85FB-E673CD5FC9E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.MongoDB", "..\Bifrost.MongoDb\Bifrost.MongoDB.csproj", "{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.Ninject.Specs", "..\Bifrost.Ninject.Specs\Bifrost.Ninject.Specs.csproj", "{0F449F28-D18B-4D52-B444-A18546F82FE6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.JavaScript", "..\Bifrost.JavaScript\Bifrost.JavaScript.csproj", "{15567D0A-4DAF-49CD-9F8F-4FA94AD889D3}"
@@ -108,6 +106,10 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.Events.RabbitMQ", "..\Bifrost.Events.RabbitMQ\Bifrost.Events.RabbitMQ.csproj", "{E004883F-E9B8-4660-B213-C764DAC7B8F3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.Events.Redis", "..\Bifrost.Events.Redis\Bifrost.Events.Redis.csproj", "{589A0F70-A043-45B3-B97E-34698A029EDB}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Read", "Read", "{5097934F-C4DE-4FC3-A0CD-57B43E834E61}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bifrost.Read.MongoDB", "..\Bifrost.Read.MongoDb\Bifrost.Read.MongoDB.csproj", "{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -207,20 +209,6 @@ Global
 		{3EE03158-1BE0-470C-85FB-E673CD5FC9E1}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{3EE03158-1BE0-470C-85FB-E673CD5FC9E1}.Release|x64.ActiveCfg = Release|Any CPU
 		{3EE03158-1BE0-470C-85FB-E673CD5FC9E1}.Release|x86.ActiveCfg = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|ARM.ActiveCfg = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x64.ActiveCfg = Release|Any CPU
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x86.ActiveCfg = Release|Any CPU
 		{0F449F28-D18B-4D52-B444-A18546F82FE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0F449F28-D18B-4D52-B444-A18546F82FE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0F449F28-D18B-4D52-B444-A18546F82FE6}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -646,6 +634,26 @@ Global
 		{589A0F70-A043-45B3-B97E-34698A029EDB}.Release|x64.Build.0 = Release|Any CPU
 		{589A0F70-A043-45B3-B97E-34698A029EDB}.Release|x86.ActiveCfg = Release|Any CPU
 		{589A0F70-A043-45B3-B97E-34698A029EDB}.Release|x86.Build.0 = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|ARM.Build.0 = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x64.Build.0 = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Debug|x86.Build.0 = Debug|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|ARM.ActiveCfg = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|ARM.Build.0 = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x64.ActiveCfg = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x64.Build.0 = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x86.ActiveCfg = Release|Any CPU
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -657,7 +665,6 @@ Global
 		{FFE4E208-67F2-47B5-AA92-BA35AB3F4730} = {2B61FCED-42AC-48B3-B2E9-6F0EE50E0674}
 		{DB77EB6A-DB79-4839-AB30-90593BCE470F} = {2B61FCED-42AC-48B3-B2E9-6F0EE50E0674}
 		{3EE03158-1BE0-470C-85FB-E673CD5FC9E1} = {1287AD93-5279-46B3-AE57-1CCC468CB2B1}
-		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA} = {5BD443D5-1CDE-4FF5-BFA7-9761EC14FE0C}
 		{0F449F28-D18B-4D52-B444-A18546F82FE6} = {1287AD93-5279-46B3-AE57-1CCC468CB2B1}
 		{15567D0A-4DAF-49CD-9F8F-4FA94AD889D3} = {C639794D-C022-4118-902C-5F003DFA0B31}
 		{F3E029DC-EBEB-492F-9CA0-D26DB9F82064} = {C639794D-C022-4118-902C-5F003DFA0B31}
@@ -690,6 +697,8 @@ Global
 		{D184DA7F-505F-45F5-A192-728705639715} = {6B7FCD5E-7B6C-44AA-A370-6EEE78D9EBAA}
 		{E004883F-E9B8-4660-B213-C764DAC7B8F3} = {6B7FCD5E-7B6C-44AA-A370-6EEE78D9EBAA}
 		{589A0F70-A043-45B3-B97E-34698A029EDB} = {6B7FCD5E-7B6C-44AA-A370-6EEE78D9EBAA}
+		{5097934F-C4DE-4FC3-A0CD-57B43E834E61} = {2B61FCED-42AC-48B3-B2E9-6F0EE50E0674}
+		{63BBC1F3-E5E0-4902-BADF-1589F91C0BEA} = {5097934F-C4DE-4FC3-A0CD-57B43E834E61}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		EnterpriseLibraryConfigurationToolBinariesPath = packages\Unity.2.1.505.0\lib\NET35;packages\Unity.2.1.505.2\lib\NET35
